fix: :bug: wrong code block undo/redo position
